agora/main_ui_backend_api/src/requirements.txt
fastapi 
uvicorn
psycopg2-binary
